refactor: first steps to deploy endpoints
