+++ .github/workflows/update.yml
@@ -1,0 +1,26 @@
+name: Update Endpoints
+
+on:
+  push:
+    branches:
+      - main
+  schedule:
+    - cron:  '0 8 * * *'
+
+permissions: 
+  contents: write
+
+jobs:
+  update:
+    runs-on: ubuntu-latest
+    steps:
+    - uses: actions/checkout@v3
+    - uses: actions/setup-node@v3
+      with:
+        node-version: 18
+    - run: npm i
+    - run: npm run endpoints
+    - uses: JamesIves/github-pages-deploy-action@v4.3.3
+      with:
+        branch: gh-pages
+        folder: data
