+++ package.json
@@ -9,8 +9,9 @@
   },
   "scripts": {
     "lint": "npm run endpoints && standard *.js bin/* && dtslint --localTs node_modules/typescript/lib types",
-    "endpoints": "node endpoints-json.js",
-    "prepack": "npm run endpoints",
+    "endpoints": "node bin/update-endpoints",
+    "build": "esm2umd",
+    "prepare": "npm run build && npm run endpoints",
     "test": "node test/env npm run test-impl",
     "test-impl": "fresh-tape test/all.js",
     "test-cov": "node test/env c8 -r lcov -r html npm run test-impl",
@@ -39,25 +40,26 @@
   "author": "Martin Heidegger <martin.heidegger@gmail.com>",
   "license": "MIT",
   "dependencies": {
-    "@leichtgewicht/ip-codec": "^2.0.2",
-    "@types/dns-packet": "^5.2.0",
-    "dns-packet": "^5.3.0",
-    "dns-socket": "^4.2.2"
+    "@leichtgewicht/dns-packet": "^5.3.3",
+    "@leichtgewicht/dns-socket": "^4.2.2",
+    "@leichtgewicht/ip-codec": "^2.0.4"
   },
   "devDependencies": {
+    "@definitelytyped/dtslint": "0.0.112",
+    "@leichtgewicht/dnsstamp": "^1.1.4",
+    "@leichtgewicht/esm2umd": "^0.1.1",
     "abort-controller": "^3.0.0",
     "browserify": "^17.0.0",
-    "c8": "^7.7.3",
+    "c8": "^7.11.3",
     "coverify": "^1.5.1",
-    "dtslint": "^4.1.1",
     "envify": "^4.1.0",
-    "fresh-tape": "^5.2.4",
-    "markdown-it": "^12.1.0",
-    "ngrok": "^4.0.1",
-    "p-map": "^4.0.0",
-    "standard": "^16.0.3",
-    "tape-run": "^9.0.0",
-    "typescript": "^4.3.5",
+    "fresh-tape": "^5.5.3",
+    "markdown-it": "^13.0.1",
+    "ngrok": "^4.3.1",
+    "p-map": "^5.4.0",
+    "standard": "^17.0.0",
+    "tape-run": "^10.0.0",
+    "typescript": "^4.6.4",
     "xhr2": "^0.2.1"
   },
   "repository": {
